Move backing configuration related code into dedicated file
