+++ Sources/ZIPFoundation/Archive+BackingConfiguration.swift
@@ -1,0 +1,125 @@
+//
+//  Archive+BackingConfiguration.swift
+//  ZIPFoundation
+//
+//  Created by Thomas Zoechling on 19.09.21.
+//
+
+import Foundation
+
+extension Archive {
+
+    struct BackingConfiguration {
+        let file: UnsafeMutablePointer<FILE>
+        let endOfCentralDirectoryRecord: EndOfCentralDirectoryRecord
+        let zip64EndOfCentralDirectory: ZIP64EndOfCentralDirectory?
+        #if swift(>=5.0)
+        let memoryFile: MemoryFile?
+
+        init(file: UnsafeMutablePointer<FILE>,
+             endOfCentralDirectoryRecord: EndOfCentralDirectoryRecord,
+             zip64EndOfCentralDirectory: ZIP64EndOfCentralDirectory? = nil,
+             memoryFile: MemoryFile? = nil) {
+            self.file = file
+            self.endOfCentralDirectoryRecord = endOfCentralDirectoryRecord
+            self.zip64EndOfCentralDirectory = zip64EndOfCentralDirectory
+            self.memoryFile = memoryFile
+        }
+        #else
+
+        init(file: UnsafeMutablePointer<FILE>,
+             endOfCentralDirectoryRecord: EndOfCentralDirectoryRecord,
+             zip64EndOfCentralDirectory: ZIP64EndOfCentralDirectory?) {
+            self.file = file
+            self.endOfCentralDirectoryRecord = endOfCentralDirectoryRecord
+            self.zip64EndOfCentralDirectory = zip64EndOfCentralDirectory
+        }
+        #endif
+    }
+
+    static func makeBackingConfiguration(for url: URL, mode: AccessMode)
+    -> BackingConfiguration? {
+        let fileManager = FileManager()
+        switch mode {
+        case .read:
+            let fileSystemRepresentation = fileManager.fileSystemRepresentation(withPath: url.path)
+            guard let archiveFile = fopen(fileSystemRepresentation, "rb"),
+                  let (eocdRecord, zip64EOCD) = Archive.scanForEndOfCentralDirectoryRecord(in: archiveFile) else {
+                return nil
+            }
+            return BackingConfiguration(file: archiveFile,
+                                        endOfCentralDirectoryRecord: eocdRecord,
+                                        zip64EndOfCentralDirectory: zip64EOCD)
+        case .create:
+            let endOfCentralDirectoryRecord = EndOfCentralDirectoryRecord(numberOfDisk: 0, numberOfDiskStart: 0,
+                                                                          totalNumberOfEntriesOnDisk: 0,
+                                                                          totalNumberOfEntriesInCentralDirectory: 0,
+                                                                          sizeOfCentralDirectory: 0,
+                                                                          offsetToStartOfCentralDirectory: 0,
+                                                                          zipFileCommentLength: 0,
+                                                                          zipFileCommentData: Data())
+            do {
+                try endOfCentralDirectoryRecord.data.write(to: url, options: .withoutOverwriting)
+            } catch { return nil }
+            fallthrough
+        case .update:
+            let fileSystemRepresentation = fileManager.fileSystemRepresentation(withPath: url.path)
+            guard let archiveFile = fopen(fileSystemRepresentation, "rb+"),
+                  let (eocdRecord, zip64EOCD) = Archive.scanForEndOfCentralDirectoryRecord(in: archiveFile) else {
+                return nil
+            }
+            fseeko(archiveFile, 0, SEEK_SET)
+            return BackingConfiguration(file: archiveFile,
+                                        endOfCentralDirectoryRecord: eocdRecord,
+                                        zip64EndOfCentralDirectory: zip64EOCD)
+        }
+    }
+
+    #if swift(>=5.0)
+    static func makeBackingConfiguration(for data: Data, mode: AccessMode)
+    -> BackingConfiguration? {
+        let posixMode: String
+        switch mode {
+        case .read: posixMode = "rb"
+        case .create: posixMode = "wb+"
+        case .update: posixMode = "rb+"
+        }
+        let memoryFile = MemoryFile(data: data)
+        guard let archiveFile = memoryFile.open(mode: posixMode) else { return nil }
+
+        switch mode {
+        case .read:
+            guard let (eocdRecord, zip64EOCD) = Archive.scanForEndOfCentralDirectoryRecord(in: archiveFile) else {
+                return nil
+            }
+
+            return BackingConfiguration(file: archiveFile,
+                                        endOfCentralDirectoryRecord: eocdRecord,
+                                        zip64EndOfCentralDirectory: zip64EOCD,
+                                        memoryFile: memoryFile)
+        case .create:
+            let endOfCentralDirectoryRecord = EndOfCentralDirectoryRecord(numberOfDisk: 0, numberOfDiskStart: 0,
+                                                                          totalNumberOfEntriesOnDisk: 0,
+                                                                          totalNumberOfEntriesInCentralDirectory: 0,
+                                                                          sizeOfCentralDirectory: 0,
+                                                                          offsetToStartOfCentralDirectory: 0,
+                                                                          zipFileCommentLength: 0,
+                                                                          zipFileCommentData: Data())
+            _ = endOfCentralDirectoryRecord.data.withUnsafeBytes { (buffer: UnsafeRawBufferPointer) in
+                fwrite(buffer.baseAddress, buffer.count, 1, archiveFile) // Errors handled during read
+            }
+            fallthrough
+        case .update:
+            guard let (eocdRecord, zip64EOCD) = Archive.scanForEndOfCentralDirectoryRecord(in: archiveFile) else {
+                return nil
+            }
+
+            fseeko(archiveFile, 0, SEEK_SET)
+            return BackingConfiguration(file: archiveFile,
+                                        endOfCentralDirectoryRecord: eocdRecord,
+                                        zip64EndOfCentralDirectory: zip64EOCD,
+                                        memoryFile: memoryFile)
+        }
+    }
+    #endif
+}
